refactor(spi): move register structures into separate register module
